Create hook to get single artist
